(build) Add blue-tape for testing
- Added Blue-Tape npm module
- Added React and Reflux npm modules for testing
- Wrote sample tests (they don't do anything special, but log the
    rendered shallow dom elements to console in response to the
    following command: babel-node <path to test.js file>). The results
    of this command can be piped to faucet
    (babel-node <path to test.js file> | faucet)
- Extracted the App component from main.jsx.js and moved it in a
  separate file for testability
